+++ packages/apollo-link-http/src/httpLink.ts
@@ -7,38 +7,63 @@ import { ApolloFetch } from 'apollo-fetch';
 // XXX replace with actual typings when available
 declare var AbortController: any;
 
-type ResponseError = Error & {
-  response?: Response;
+//Used for any Error for data from the server
+//on a request with a Status >= 300
+//response contains no data or errors
+type ServerError = Error & {
+  response: Response;
+  result: Record<string, any>;
+  statusCode: number;
+};
+
+//Thrown when server's resonse is cannot be parsed
+type ServerParseError = Error & {
+  response: Response;
+  statusCode: number;
+};
+
+type ClientParseError = Error & {
   parseError: Error;
-  statusCode?: number;
+};
+
+const throwServerError = (response, result, message) => {
+  const error = new Error(message) as ServerError;
+
+  error.response = response;
+  error.statusCode = response.status;
+  error.result = result;
+
+  throw error;
 };
 
 const parseAndCheckResponse = request => (response: Response) => {
   return response
     .json()
+    .catch(e => {
+      const parseError = e as ServerParseError;
+      parseError.response = response;
+      parseError.statusCode = response.status;
+
+      throw parseError;
+    })
     .then(result => {
-      if (response.status >= 300)
-        throw new Error(
+      if (response.status >= 300) {
+        //Network error
+        throwServerError(
+          response,
+          result,
           `Response not successful: Received status code ${response.status}`,
         );
+      }
       if (!result.hasOwnProperty('data') && !result.hasOwnProperty('errors')) {
-        throw new Error(
+        //Data error
+        throwServerError(
+          response,
+          result,
           `Server response was missing for query '${request.operationName}'.`,
         );
       }
       return result;
-    })
-    .catch(e => {
-      const httpError = new Error(
-        `Network request failed with status ${response.status} - "${
-          response.statusText
-        }"`,
-      ) as ResponseError;
-      httpError.response = response;
-      httpError.parseError = e;
-      httpError.statusCode = response.status;
-
-      throw httpError;
     });
 };
 
@@ -145,7 +170,7 @@ export const createHttpLink = (
         } catch (e) {
           const parseError = new Error(
             `Network request failed. Payload is not serializable: ${e.message}`,
-          ) as ResponseError;
+          ) as ClientParseError;
           parseError.parseError = e;
           throw parseError;
         }
@@ -153,7 +178,7 @@ export const createHttpLink = (
         let options = fetchOptions;
         if (requestOptions.fetchOptions)
           options = { ...requestOptions.fetchOptions, ...options };
-        const fetcherOptions = {
+        const fetcherOptions: any = {
           method: 'POST',
           ...options,
           headers: {
@@ -219,7 +244,7 @@ export const createHttpLink = (
 
 export class HttpLink extends ApolloLink {
   public requester: RequestHandler;
-  constructor(opts: FetchOptions) {
+  constructor(opts?: FetchOptions) {
     super(createHttpLink(opts).request);
   }
 }