+++ packages/apollo-link/src/index.ts
@@ -1,5 +1,5 @@
 export * from './link';
-export { makePromise } from './linkUtils';
+export { makePromise, toPromise, fromPromise } from './linkUtils';
 export * from './types';
 
 import * as Observable from 'zen-observable';